+++ ChangeLog.rst
@@ -1,6 +1,11 @@
 Unreleased Changes
 ==================
 
+* There are now new `fuse_session_unmount` and `fuse_session_mount`
+  functions that should be used in the low-level API. The
+  `fuse_mount` and `fuse_unmount` functions should be used with the
+  high-level API only.
+
 * The ``fuse_lowlevel_notify_*`` functions now all take a `struct
   fuse_session` parameter instead of a `struct fuse_chan`.